Wrote a byte array parser for SID messages
The parser class is called SidByteParser. Also wrote an exception class,
SidByteParserException, and added a reference to a third-party library
called MiscUtil, which the parser uses.

Made use of the parser in the only concrete message class so far:
AuthInfoServerToClientSidMessage.

Changed the underlying type of the LogonType enumeration from uint to int.