Set up the unit test project with NUnit
The NUnit framework assembly is required to compile the unit test project.
This should be managed automatically by NuGet when NuGet Package Restore
is enabled (right-click on the solution in the Solution Explorer and click
"Enable NuGet Package Restore" to enable it).

To run NUnit tests in Visual Studio 2012, the NUnit Test Adapter is
required. It is available as an extension to Visual Studio 2012, though
not to the express edition ("Extensions and Updates" in the Tools menu).
The test adapter is also available as a NuGet package, which would be
preferable, but the package did not seem to work on Visual Studio 2012
Express (Update 3), so it is not included in the repository yet.