Added some contracts to the SidByteParser class
Added some ensures contracts to the methods and properties of the
SidByteParser class. Also added some object invariants for the class.
These contracts are checked by the runtime checker, and they are
nonintrusive in the sense that the project can be built with the runtime
disabled and possibly even without Code Contracts installed (without
running ccrewrite on the assembly, to be specific).

The static checker seems to be flawed and hard to work with (and requires
the Premium edition (or better) of Visual Studio), so it is disabled.