Wrote a C>S SID_PING class with some unit tests
The two SID_PING classes are essentially the same (find and replace client
with server and vice versa). This might be okay, but the boilerplate code
shared by all SID messages should probably be reduced at the expense of
simplicity.

There will probably be similar problems with the corresponding unit test
fixtures.